moved cardflip into components
